requirements.txt
jupyter
chardet
pyyaml
requests
pytest
pytest-cov
fabric3
sqlalchemy
pycryptodome
sphinx